pyproject.toml
[project]
name = "flippy"
version = "0.1.0"
description = "Hanover Flip-Dot Controller"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4~=4.12.2",
    "blessed~=1.20.0",
    "numpy~=1.26.1",
    "pillow~=10.3.0",
    "pyserial~=3.5",
    "python-dotenv~=1.0.0",
    "requests~=2.32.0",
    "spotipy~=2.25.1",
    "unidecode~=1.3.7",
]

[project.scripts]
flippy_demo = "demo.__main__:run_demo"

[tool.uv]
package = true

[tool.setuptools]
packages = ["demo", "flippy", "utils"]